MinervaV0/Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.4
boto3==1.34.0
Pillow==10.1.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1